Move ExtractCustomServiceConfig to utils.go
