+++ pkg/servicecfg/edpm.go
@@ -170,5 +170,4 @@ func CompareMappingConfig(srcMap map[string]string, configMapping map[string]str
 	}
 	godiff.PrintReport(report)
 	return nil
-
 }
